+++ ldap/servers/slapd/back-ldbm/db-mdb/mdb_layer.c
@@ -811,7 +811,7 @@ dbmdb_copyfile(char *source, char *destination, int overwrite __attribute__((unu
     int dest_fd = -1;
     char *buffer = NULL;
     int return_value = -1;
-    int bytes_to_write = 0;
+    size_t bytes_to_write = 0;
 
     /* malloc the buffer */
     buffer = slapi_ch_malloc(64 * 1024);
